backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-butler"
version = "1.0.0"
description = "FastAPI backend for managing AI prompts with markdown frontmatter storage"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.1",
    "pyyaml==6.0.2",
    "pydantic==2.10.3",
    "python-multipart==0.0.12",
    "python-dotenv==1.0.1",
    "python-frontmatter>=1.1.0",
    "rapidfuzz>=3.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyperclip>=1.8.0",
    "textual>=3.0.0",
]

[project.scripts]
pb = "prompt_butler.cli:app"
pb-tui = "prompt_butler.tui.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
]

[tool.setuptools]
packages = [
    "prompt_butler",
    "prompt_butler.services",
    "prompt_butler.routers",
    "prompt_butler.tui",
]

[tool.ruff]
line-length = 120
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
